third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/shadow-dom-mouse-event-adjust-offset-expected.txt

CONSOLE ERROR: line 87: Uncaught TypeError: element.createShadowRoot is not a function
Tests to ensure that MouseEvent's offsetX and offsetY are adjusted in re-targeting.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
